[ineco] fix review comments
